REFACTOR: Relocate and documents types/functions for model validation.
