+++ packages/frontend/src/model/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
+export * from "./context";
 export * from "./object_cell_editor";
 export * from "./morphism_cell_editor";
